requirements.txt
slacker
python-dotenv